changing how validator snapshots are retrieved and better logging
as we failed to submit some checkpoints whilst a new validator was added to the network we cannot send old checkpoints any more. This is because the latest validator snapshot is being returned from the call to cache which has a new validator in it that the L1 contract doesn't know about.